pyproject.toml
[project]
name = "playwright-prueba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "dotenv>=0.9.9",
    "playwright>=1.52.0",
    "polars>=1.30.0",
    "psycopg2-binary>=2.9.10",
    "pytest-playwright>=0.7.0",
    "sqlalchemy>=2.0.41",
    "tenacity>=9.1.2",
]
